backend/requirements.txt
# Django核心
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
django-filter==23.3

# 数据库
mysqlclient==2.2.0
pymysql==1.1.0

# Redis缓存
django-redis==5.4.0
redis==5.0.1

# 异步任务
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1

# JWT认证
djangorestframework-simplejwt==5.3.0

# API文档
drf-yasg==1.21.7

# 工具库
python-dotenv==1.0.0
Pillow==10.1.0
markdown==3.5.1
pygments==2.17.2
requests==2.31.0

# 开发工具
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# 部署相关
gunicorn==21.2.0
whitenoise==6.6.0
